requirements.txt
# Project DiMaggio — Dependencies

# Data & ML
pybaseball>=2.2.7
duckdb>=0.10.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.4.0
torch>=2.2.0
torchvision>=0.17.0

# Reinforcement Learning
gymnasium>=0.29.1
stable-baselines3>=2.2.1

# Explainability
shap>=0.44.0

# Progress bars
tqdm>=4.66.0

# API
fastapi>=0.110.0
uvicorn[standard]>=0.28.0
pydantic>=2.6.0

# Weather
requests>=2.31.0

# Dashboard
streamlit>=1.32.0
plotly>=5.20.0

# Utilities
python-dotenv>=1.0.1
pyarrow>=15.0.0
tqdm>=4.66.0
rich>=13.7.0
click>=8.1.7

# Testing
pytest>=8.0.0
httpx>=0.27.0  # for FastAPI TestClient
